Build enhancements to support easier change tracking (#1424)
Will be adding a monorepo which contains the edge-* repos as git
submodules to track changes across nightly builds better, especially
when branches are involved. The monorepo will be used only for nightly
builds ,and will automatically pull in and tag the changes from each of
the submodules.

With a monorepo, we need the top-level ".git" directory to be included
in the docker build context. Hence this change.

Note: The monorepo won't change anything of how people build the
      workspaces manually currently.