EDGECLOUD-5790 App instance health status wrongly shows “Cloudletoffline” status for an online & operational cloudlet  (#1551)
We discovered that we have some appInstances in the CloudletOffline state, even though the cloudlet is online. This is due to the fact that we don't properly clean up appInst state when cloudlet offline alert is cleaned up.
When we delete cloudlet offline alert we should also set the health check to OK for each appInst on the cloudlet that's back online.